Refactor heads-up display
